Final Camera Project: Implement Midterm Project Code
This commit implements the keypoint detector/descriptor code which was
developed in the midterm project.

It allows the user to select different detectors and descriptors to match
keypoints on the provided images. The different possibilities are commented
with the variable.

Also it is possible to switch between the Bruteforce and the FLANN matching.