Final Camera Project: Implement Final Project Code
This commit introduces the fusion between the lidar data and the camera
images.

The distance for the lidar was calculated by the mean distance of the
x values from the measurement.

Based on the bounding boxes found by the code from the midterm project and
the corresponding matches. The box is tracked over multiple frames.

Based on the mean distance of  the matches the ratio was calculated and
the Time to collision for the camera was calculated.

To gain an understanding of how long each step took there is the time used
per image and sensor at the end of the program.